+++ templates/home.html
@@ -43,14 +43,6 @@
 				<div class="logo">
 					<h1>What Slot</h1>
 				</div>
-				<a href="https://github.com/metakgp/what-slot/"> <i data-feather='github'></i> </a>
-				<i data-feather='sun'></i>
-				<label class="switch">
-					<input type="checkbox" onclick="toggleDarkLight()">
-					<span class="slider round"></span>
-				</label>
-				<i data-feather='moon'></i>
-
 			</nav>
 
 			<div class="timetable">
